Added html  pages and urls in notes
Created an html page in templates for the notes app, and added a urls file since it did not exist. Connect the template pages in the urls. After that, include the notes url path in the smartnotes url pattern, which is the main project. Any app added with a url, the path needs to be included in the project's urls, and any app added needs to be mentioned in the project's settings under installed apps.