The D in CRUD: Deleting Data
Deleted the data in the database and on the web as well. These things are connected in a way. I haven't confirmed it yet that the data is removed from the database, I'm just seeing that the db.sqlite3 file has been modified after I have created, update or deleted a note.