Adding a new note after ForeignKey
Added a new note with the user logged in. Before calling the save method we set the commit to false so that it doesn't save to the database. Saving to the database without setting the user will cause an error stating that "You can't have a null user". The form_valid method is overridden so that the process of saving to the database can be interjected so that we can inject the user of the note being created.