Added login and logout pages
Created the login and logout page. With the login html you are creating a form that has a post.. You will need to import LoginView and LogoutView to be able to access the functionality of a login and of a logout.  Add a "LOGIN_REDIRECT_URL" in settings to set a path on where to go after logging in to the system.